vit_emotion_api/requirements.txt
fastapi
uvicorn
torch
transformers
opencv-python
pillow
